fixing some  issues on implementing ML
